AMDb Recommendation System/requirements.txt
asgiref==3.6.0
astroid==2.14.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
cymem==2.0.7
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
Django==4.1.6
django-crispy-forms==1.14.0
django-environ==0.9.0
django-import-export==3.0.2
et-xmlfile==1.1.0
idna==3.4
isort==5.12.0
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
lazy-object-proxy==1.9.0
MarkupPy==1.14
MarkupSafe==2.1.2
mccabe==0.7.0
murmurhash==1.0.9
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.0
packaging==23.0
pathy==0.10.1
platformdirs==3.0.0
preshed==3.0.8
pydantic==1.10.5
pylint==2.16.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.5
tablib==3.3.0
thinc==8.1.7
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.14
wasabi==1.1.1
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0